Speed tweak to sequence parser
Fasta files are chock full of newlines.  I caused errors when using this
library in BamBayes because I forgot to remove the newlines.  I
attempted to remove them with boost::erase, but that was amazingly slow.
A better solution is to use std::remove (it pushes the element to remove
to the end of the string), then resize the string to truncate the
unwanted characters.