remove boost dependency, add split, bugfix
To remove boost dependency, use stl functions and c stdlib functions,
add functions for splitting a string into a vector on a set of
delimiters (Split.cpp).  Fixes a bug with getSequence, that allowed
non-null-terminated strings to be passed through, and a bug with
getSubSequence which returned strings which had long strings of trailing
nulls in the case that the subsequence overlapped the end of the
sequence.