add stdin region parsing
You can now pass a stream of region specifiers to fastahack, and it will
print the sequence on its stdout.  This is intended to simplify use in
dynamic languages until I generate a SWIG binding.

Additionally... fastahack now uses a 1-base start 1-base end region
string with the end *included* in the region to return.